merge develop -> master

+++ .build/create_repository.py
@@ -11,6 +11,12 @@ import requests
 import zipfile
 import glob
 from distutils.version import LooseVersion
+import mistune
+import sys
+
+# sys.setdefaultencoding() does not exist, here!
+reload(sys)  # Reload does the trick!
+sys.setdefaultencoding('UTF8')
 
 plugins_dir = 'addons'
 build_dir = 'build'
@@ -23,7 +29,7 @@ plugins_info = config['plugins']
 repository = config['repository']
 repo_name = repository['id']
 repo_url = repository['url']
-repo_github_url = repository['github_url']
+repo_source_url = repository['source_url']
 
 
 def init():
@@ -42,14 +48,14 @@ def build_plugins():
 
 	for plugin_info in plugins_info:
 		name = plugin_info['name']
-		print name
-		github_url = plugin_info['github_url']
+		print "processing: %s" % name
+		source_url = plugin_info['source_url']
 
-		if github_url.endswith('.git'):
-			clone_url = github_url
-			github_url = github_url[:-4]
+		if source_url.endswith('.git'):
+			clone_url = source_url
+			source_url = source_url[:-4]
 		else:
-			clone_url = "%s.git" % github_url
+			clone_url = "%s.git" % source_url
 
 		repo_dir = os.path.join(plugins_dir, name)
 		if not os.path.isdir(repo_dir):
@@ -69,37 +75,27 @@ def build_plugins():
 		if len(tag_list) == 0:
 			# repo doesn't have any tags, so we will generate the repo from the 'latest'
 			# we will get the version from the addon.xml
-			print "no tags"
-			_process_non_tagged_addon(repo, name, addons_xml_root)
+			_process_non_tagged_addon(repo, plugin_info, addons_xml_root)
 		else:
 
 			latest_processed = False
 			for tag in tag_list:
 
-				print "tag: %s" % tag
-
 				version = _get_version_from_tag(name, tag)
-				print "version: %s" % version
-
 				if _is_tag_filtered_out(plugin_info, tag):
 					continue
 
 				repo.git.checkout('tags/%s' % tag)
 
-				changelog_file = os.path.join(repo_dir, "changelog.txt")
-				if not os.path.isfile(changelog_file):
-					print "skipping: %s - No changelog" % version
-					continue
-
 				name_with_version = '%s-%s' % (name, version)
-				print name_with_version
 				# try to download pre-existing zip from github releases
+
 				try:
-					if _process_release_addon(tag, plugin_info, addons_xml_root, latest_processed):
+					if _process_github_release_addon(repo_dir, tag, plugin_info, addons_xml_root, latest_processed):
 						latest_processed = True
 				except Exception as err:
 					print "download failed: {0}".format(err)
-					if _process_non_release_addon(repo_dir, name, version, addons_xml_root, latest_processed):
+					if _process_non_release_addon(repo_dir, plugin_info, version, addons_xml_root, latest_processed):
 						latest_processed = True
 
 	# remove temp path
@@ -113,12 +109,18 @@ def build_plugins():
 	_md5_hash_file(addon_xml_file)
 
 
-def _process_release_addon(tag, plugin_info, addons_xml_root, latest_processed):
+def _process_github_release_addon(repo_dir, tag, plugin_info, addons_xml_root, latest_processed):
 	name = plugin_info["name"]
 	version = _get_version_from_tag(name, tag)
-	github_url = plugin_info["github_url"]
-	if github_url.endswith('.git'):
-		github_url = github_url[:-4]
+	source_url = plugin_info["source_url"]
+
+	if not source_url.startswith("http://github.com/") and not source_url.startswith("https://github.com/"):
+		# only github releases are currently supported.
+		# if it is not github, then don't even try to get it, it will just fail.
+		return _process_non_release_addon(repo_dir, name, version, addons_xml_root, latest_processed)
+
+	if source_url.endswith('.git'):
+		source_url = source_url[:-4]
 	addons_xml_added = False
 	name_with_version = "%s-%s" % (name, version)
 
@@ -130,14 +132,14 @@ def _process_release_addon(tag, plugin_info, addons_xml_root, latest_processed):
 		os.mkdir(temp_extract_path)
 
 	try:
-		release_zip_url = '%s/releases/download/%s/%s.zip' % (github_url, tag, name_with_version)
+		release_zip_url = '%s/releases/download/%s/%s.zip' % (source_url, tag, name_with_version)
 		print release_zip_url
 		local_filename = _download_file(release_zip_url, build_temp_dir)
 		with zipfile.ZipFile(local_filename, 'r') as zip_ref:
 			zip_ref.extractall(temp_extract_path)
 	except Exception as oops:
 		# some break the rules and put the v in the package... lets fix it.
-		release_zip_url = '%s/releases/download/%s/%s-v%s.zip' % (github_url, tag, name, version)
+		release_zip_url = '%s/releases/download/%s/%s-v%s.zip' % (source_url, tag, name, version)
 		local_filename = _download_file(release_zip_url, os.path.join(build_temp_dir, "%s.zip" % name_with_version))
 		with zipfile.ZipFile(local_filename, 'r') as zip_ref:
 			zip_ref.extractall(temp_extract_path)
@@ -157,6 +159,11 @@ def _process_release_addon(tag, plugin_info, addons_xml_root, latest_processed):
 
 	_md5_hash_file(version_zip)
 
+	readmes = glob.glob1(build_plugin_path, "readme*")
+	if len(readmes) > 0:
+		readme = readmes[0]
+		shutil.copy2(os.path.join(build_plugin_path, readme), os.path.join(build_plugin_version_path, "readme.md"))
+
 	if os.path.exists(os.path.join(build_plugin_version_path, 'changelog.txt')):
 		shutil.move(os.path.join(build_plugin_version_path, 'changelog.txt'),
 		            os.path.join(build_plugin_version_path, 'changelog-%s.txt' % version))
@@ -172,6 +179,11 @@ def _process_release_addon(tag, plugin_info, addons_xml_root, latest_processed):
 
 	shutil.move(version_zip, os.path.join(build_plugin_path, "%s.zip" % name_with_version))
 	shutil.move("%s.md5" % version_zip, os.path.join(build_plugin_path, "%s.zip.md5" % name_with_version))
+
+	if os.path.exists(os.path.join(build_plugin_version_path, 'readme.md')):
+		shutil.copy2(os.path.join(build_plugin_version_path, 'readme.md'),
+		             os.path.join(build_plugin_path, 'readme.md'))
+
 
 	if os.path.exists(os.path.join(build_plugin_version_path, 'changelog-%s.txt' % version)):
 		shutil.copy2(os.path.join(build_plugin_version_path, 'changelog-%s.txt' % version),
@@ -190,23 +202,27 @@ def _process_release_addon(tag, plugin_info, addons_xml_root, latest_processed):
 	return addons_xml_added
 
 
-def _process_non_tagged_addon(repo, name, addons_xml_root):
+def _process_non_tagged_addon(repo, plugin_info, addons_xml_root):
+	name = plugin_info["name"]
 	repo_dir = os.path.join(plugins_dir, name)
+	plugin_info_path = plugin_info["path"] if "path" in plugin_info else ""
+	repo_plugin_dir = os.path.join(repo_dir, plugin_info_path)
 	repo.git.checkout("HEAD")
 
-	if not _can_process_non_tagged(repo_dir):
+	if not _can_process_non_tagged(repo_plugin_dir):
 		print "unable to process because required files are missing"
 		return
 
-	plugin_addon_xml = etree.parse(open(os.path.join(repo_dir, 'addon.xml')))
+	plugin_addon_xml = etree.parse(open(os.path.join(repo_plugin_dir, 'addon.xml')))
 	version = _get_version_from_addon_tree(plugin_addon_xml)
 
-	_process_non_release_addon(repo_dir, name, version, addons_xml_root, False)
+	_process_non_release_addon(repo_dir, plugin_info, version, addons_xml_root, False)
+
 
 # checks for the 'required' files to process
 def _can_process_non_tagged(repo_dir):
-	if not os.path.isfile(os.path.join(repo_dir, "changelog.txt")):
-		return False
+	# if not os.path.isfile(os.path.join(repo_dir, "changelog.txt")):
+	# 	return False
 	if not os.path.isfile(os.path.join(repo_dir, "icon.png")):
 		return False
 	if not os.path.isfile(os.path.join(repo_dir, "addon.xml")):
@@ -218,7 +234,8 @@ def _get_version_from_addon_tree(addon_tree):
 	return addon_tree.getroot().attrib["version"]
 
 
-def _process_non_release_addon(repo_dir, name, version, addons_xml_root, latest_processed):
+def _process_non_release_addon(repo_dir, plugin_info, version, addons_xml_root, latest_processed):
+	name = plugin_info["name"]
 	addons_xml_added = False
 	name_with_version = "%s-%s" % (name, version)
 	temp_extract_path = build_temp_dir
@@ -228,7 +245,15 @@ def _process_non_release_addon(repo_dir, name, version, addons_xml_root, latest_
 		os.remove(os.path.join(build_temp_dir, "%s.zip" % name_with_version))
 	build_repo_path = os.path.join(temp_extract_path, name)
 	# copy the git repo to build_repo_path
-	shutil.copytree(repo_dir, build_repo_path, ignore=shutil.ignore_patterns('.git*'))
+	plugin_info_path = plugin_info["path"] if "path" in plugin_info else ""
+
+	joined_repo_path = os.path.join(repo_dir, plugin_info_path)
+	shutil.copytree(joined_repo_path, build_repo_path, ignore=shutil.ignore_patterns('.git*'))
+
+	if len(glob.glob1(build_repo_path, "readme*")) < 1:
+		treadmes = glob.glob1(repo_dir, "readme*")
+		if len(treadmes) > 0:
+			shutil.copy2(os.path.join(repo_dir, treadmes[0]), os.path.join(build_repo_path, "readme.md"))
 
 	plugin_addon_xml = etree.parse(open(os.path.join(build_repo_path, 'addon.xml')))
 	real_version = _get_version_from_addon_tree(plugin_addon_xml)
@@ -254,8 +279,14 @@ def _process_non_release_addon(repo_dir, name, version, addons_xml_root, latest_
 		if os.path.exists(l):
 			os.remove(l)
 
-	shutil.move(os.path.join(build_repo_path, 'changelog.txt'),
-	            os.path.join(build_plugin_path, 'changelog-%s.txt' % version))
+	if os.path.exists(os.path.join(build_repo_path, 'changelog.txt')):
+		shutil.move(os.path.join(build_repo_path, 'changelog.txt'),
+		            os.path.join(build_plugin_path, 'changelog-%s.txt' % version))
+
+	readmes = glob.glob1(build_repo_path, "readme*")
+	if len(readmes) > 0:
+		readme = readmes[0]
+		shutil.copy2(os.path.join(build_repo_path, readme), os.path.join(build_plugin_path, "readme.md"))
 
 	if os.path.exists(os.path.join(build_repo_path, "fanart.jpg")):
 		shutil.move(os.path.join(build_repo_path, "fanart.jpg"), latest_fanart)
@@ -312,6 +343,14 @@ def _md5_hash_file(file_name):
 		f.write(hash_md5.hexdigest())
 
 
+def _get_file_contents(file_name):
+	data = ""
+	with open(file_name, "rb") as f:
+		for chunk in iter(lambda: f.read(4096), b""):
+			data += unicode(chunk, 'utf-8')
+	return data
+
+
 def _download_file(url, dest):
 	if os.path.isdir(dest):
 		local_filename = os.path.join(dest, url.split('/')[-1])
@@ -328,7 +367,7 @@ def _download_file(url, dest):
 def _cleanup_path(path):
 	for f in os.listdir(path):
 		if f.startswith('changelog') or f.startswith('fanart.') or f.startswith('icon.') or \
-				f.endswith('.zip'):
+				f.endswith('.zip') or f.startswith("readme") or f.startswith("README"):
 			pass
 		else:
 			_f = os.path.join(path, f)
@@ -352,6 +391,7 @@ def build_gh_pages(root, current_dir):
 	html += "<a href=\"%s\">%s</a>\n" % (item, "../")
 
 	dir_items = os.listdir(cur_dir)
+	page_footer = ""
 	for item in dir_items:
 		item_path = os.path.join(cur_dir, item)
 		if os.path.isdir(item_path):
@@ -359,14 +399,27 @@ def build_gh_pages(root, current_dir):
 
 			build_gh_pages(root, os.path.join(current_dir, item))
 		else:
+			if item.startswith("readme"):
+				try:
+					page_footer = mistune.markdown(_get_file_contents(os.path.join(cur_dir, item)))\
+						.replace('&lt;', '<')\
+						.replace('&gt;', '>')
+				except UnicodeDecodeError as mde:
+					print mde.message
+					pass
+
 			html += "<a href=\"%s\">%s</a>\n" % (item, item)
 	ts = datetime.datetime.now().strftime('%d-%b-%Y %H:%M')
 	doc_version = os.getenv('CI_BUILD_VERSION', '1.0.0.0')
-	html += "</pre><hr/><div>Generated by <a href=\"%s\">%s</a> v%s at %s</div></body></html>" \
-	        % (repo_github_url, repo_name, doc_version, ts)
+	if page_footer != "":
+		pf = ("%s<hr/>" % page_footer)
+	else:
+		pf = ""
+	html += "</pre><hr/>%s<div>Generated by <a href=\"%s\">%s</a> v%s at %s</div></body></html>" \
+	        % (pf, repo_source_url, repo_name, doc_version, ts)
 
 	with open(os.path.join(cur_dir, 'index.html'), 'w') as f:
-		f.write(html)
+		f.write(unicode(html))
 
 
 def _build_gh_readme():
@@ -376,12 +429,12 @@ def _build_gh_readme():
 	for plugin_info in plugins_info:
 		name = plugin_info['name']
 
-		github_url = plugin_info['github_url']
+		source_url = plugin_info['source_url']
 
-		if github_url.endswith('.git'):
-			github_url = github_url[:-4]
+		if source_url.endswith('.git'):
+			source_url = source_url[:-4]
 		else:
-			github_url = "%s.git" % github_url
+			source_url = "%s.git" % source_url
 		pi_dir = os.path.join(build_plugins_dir, name)
 		dir_items = glob.glob1(pi_dir, "*.zip")
 		dir_items.sort(key=lambda x: LooseVersion(_get_version_from_tag(name, x)), reverse=True)
@@ -393,26 +446,26 @@ def _build_gh_readme():
 				c = {
 					"name": name,
 					"version": _get_version_from_tag(name, item),
-					"url": github_url
+					"url": source_url
 				}
 				data.append(c)
 
 	readme_file = os.path.join(build_plugins_dir, "readme.md")
 	if os.path.exists(readme_file):
 		os.remove(readme_file)
-	markdown = "# [%s](%s)\n\n|  | Plugin | Version | Download |\n| ----- | ----- | ----- | ----- |\n" \
+	md = u"# [%s](%s)\n\n|  | Plugin | Version | Download |\n| ----- | ----- | ----- | ----- |\n" \
 	           % (repo_name, repo_url)
 	for pd in data:
 		dl = "%s/%s/%s-%s.zip" % (repo_url, pd["name"], pd["name"], pd["version"])
 		md5 = "%s.md5" % dl
 		# https://github.com/camalot/repository.camalot/raw/gh-pages/plugin.image.kodur/icon.png
-		markdown += \
-			"| <img src=\"%s/raw/gh-pages/%s/icon.png\" width=\"24\" /> | [%s](%s) | %s | [[DL](%s)] [[MD5](%s)] |\n" \
-			% (repo_github_url, pd["name"], pd["name"], pd["url"], pd["version"], dl, md5)
+		md += \
+			u"| <img src=\"%s/raw/gh-pages/%s/icon.png\" width=\"24\" /> | [%s](%s) | %s | [[DL](%s)] [[MD5](%s)] |\n" \
+			% (repo_source_url, pd["name"], pd["name"], pd["url"], pd["version"], dl, md5)
 	ts = datetime.datetime.now().strftime('%d-%b-%Y %H:%M')
 	doc_version = os.getenv('CI_BUILD_VERSION', '1.0.0.0')
 
-	markdown += "----\n\nRepository version %s - %s\n\n----\n" % (doc_version, ts)
+	md += u"----\n\nRepository version %s - %s\n\n----\n" % (doc_version, ts)
 
 	with open('.ghp_readme.md', 'r') as ghtf:
 		data = ghtf.read()\
@@ -420,18 +473,18 @@ def _build_gh_readme():
 			.replace("#{REPO_NAME}", repo_name)\
 			.replace("#{REPO_VERSION}", repo_version)
 
-	markdown += "\n%s\n" % data
+	md += u"\n%s\n" % data
 
 	with open(readme_file, 'w') as f:
-		f.write(markdown)
+		f.write(md)
 
 
 @click.command()
 def run():
 	init()
 	build_plugins()
-	build_gh_pages(os.path.abspath(build_dir), '')
 	_build_gh_readme()
+	build_gh_pages(os.path.abspath(build_dir), '')
 
 
 if __name__ == "__main__":
